// app/src/generated/models/Al_failreasonsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Al_failreasonsal_category = {
  120910400: 'AML',
  120910401: 'Breach',
  120910402: 'Record Keeping',
  120910403: 'Tax check'
} as const;
export type Al_failreasonsal_category = keyof typeof Al_failreasonsal_category;
export const Al_failreasonsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Al_failreasonsstatecode = keyof typeof Al_failreasonsstatecode;
export const Al_failreasonsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Al_failreasonsstatuscode = keyof typeof Al_failreasonsstatuscode;

export interface Al_failreasonsBase {
  al_category: Al_failreasonsal_category;
  al_displayorder?: number;
  al_failreasoncode: string;
  al_failreasonid: string;
  al_name: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Al_failreasonsstatecode;
  statuscode?: Al_failreasonsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Al_failreasons extends Al_failreasonsBase {
  al_categoryname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}